pyproject.toml
[tool.poetry]
name = "jamsite"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.122.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
Jinja2 = "^3.1.3"
boto3 = "^1.34.65"
nltk = "^3.8.1"
Unidecode = "^1.3.8"
dropbox = "^11.36.2"
setuptools = "^69.2.0"
black = "^24.10.0"
gotenberg-client = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
jamsite = "jamsite.jamsite:main"
